feat(PasswordStrategy): authenticate via a username and password
The index action has been overwritten to fetch the user based on
Guardian and display if the user is logged in or not. Currently the
password is passed through the url params. The User module is hard coded
to expect "user@example.com" and "password" and will always return the
same struct.